separated std-lib solutions
